make base image for manifests configurable
also do not fail tests when no online access
