+++ lib/manifest.js
@@ -97,10 +97,12 @@ module.exports.generateManifestForRMarkdown = function (job_id, payload_dir, mai
   // dockerfile command must use the correct files
   r_dockerfile_command = 'dockerfile('
     + 'from = \'' + path_to_mainfile_in_container + '\', '
+    + 'image = \'' + config.containerit.baseImage + '\', '
     + 'maintainer = \'' + config.containerit.maintainer + '\', '
     + 'copy = NA, '
     + 'container_workdir = \'' + path_to_workdir_in_container + '\', '
-    + 'cmd = ' + r_render_command + ')';
+    + 'cmd = ' + r_render_command + ', '
+    + 'filter_baseimage_pkgs = ' + config.containerit.filterBaseImagePkgs + ')';
 
   // FIXME hack to remove .html .pdf files matching the main file
   main_file_without_extension = path.basename(main_file).replace(path.extname(main_file), '');
@@ -167,10 +169,12 @@ module.exports.generateManifestForRScript = function (job_id, payload_dir, main_
 
   r_dockerfile_command = 'dockerfile('
     + 'from = \'' + path_to_mainfile_in_container + '\', '
+    + 'image = \'' + config.containerit.baseImage + '\', '
     + 'maintainer = \'' + config.containerit.maintainer + '\', '
     + 'copy = NA, '
     + 'container_workdir = \'' + path_to_workdir_in_container + '\', '
-    + 'cmd = ' + r_render_command + ')';
+    + 'cmd = ' + r_render_command + '\, '
+    + 'filter_baseimage_pkgs = ' + config.containerit.filterBaseImagePkgs + ')';
 
   r_command = 'setwd(\'' + path_to_workdir_in_container + '\'); write(' + r_dockerfile_command + ')'
   //+ 'file = \'' + path.join(config.bagtainer.mountLocationInContainer, config.bagtainer.manifestFile) + '\')';
