+++ test/Dockerfile
@@ -1,4 +1,5 @@
-FROM rocker/r-ver:3.4.3
+# base image here must match config.containerit.baseImage
+FROM rocker/geospatial:3.4.4
 LABEL maintainer="o2r"
 RUN export DEBIAN_FRONTEND=noninteractive; apt-get -y update \
  && apt-get install -y git-core \
