+++ test/job-steps.js
@@ -521,216 +521,6 @@ describe('API job steps', () => {
     }).timeout(60000);
   });
 
-  describe('Manifest generation and image build for workspace minimal-rmd-data', () => {
-    let job_id = '';
-    let compendium_id = '';
-
-    before(function (done) {
-      this.timeout(90000);
-      createCompendiumPostRequest('./test/workspace/minimal-rmd-data', cookie_o2r, 'workspace', (req) => {
-        request(req, (err, res, body) => {
-          compendium_id = JSON.parse(body).id;
-          publishCandidate(compendium_id, cookie_o2r, () => {
-            startJob(compendium_id, id => {
-              job_id = id;
-              waitForJob(job_id, (finalStatus) => {
-                done();
-              });
-            });
-          });
-        });
-      });
-    });
-
-    it('should skip previous steps', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        assert.propertyVal(response.steps.validate_bag, 'status', 'skipped');
-        assert.propertyVal(response.steps.validate_compendium, 'status', 'success');
-        done();
-      });
-    });
-
-    it('should complete step "generate_manifest"', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-        assert.propertyVal(response.steps.generate_manifest, 'status', 'success');
-        done();
-      });
-    });
-
-    it('show have the manifest file in the job files', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/job/' + job_id + '/data/Dockerfile');
-        done();
-      });
-    });
-
-    it('should have the manifest file in the compendium files', (done) => {
-      request(global.test_host + '/api/v1/compendium/' + compendium_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/compendium/' + compendium_id + '/data/Dockerfile');
-        done();
-      });
-    });
-
-    it('should have the expected content in the manifest file via the job', function (done) {
-      request(global.test_host_transporter + '/api/v1/job/' + job_id + '/data/Dockerfile', (err, res, body) => {
-        assert.ifError(err);
-        assert.isNotObject(body, 'response is not JSON');
-        assert.include(body, 'FROM rocker/r-ver:3.4.3');
-        assert.include(body, 'rmarkdown::render(input = \\"/erc/main.Rmd\\"');
-        done();
-      });
-    });
-
-    it('should have the expected content in the manifest file via the compendium', function (done) {
-      request(global.test_host_transporter + '/api/v1/compendium/' + compendium_id + '/data/Dockerfile', (err, res, body) => {
-        assert.ifError(err);
-        assert.isNotObject(body, 'response is not JSON');
-        assert.include(body, 'FROM rocker/r-ver:3.4.3');
-        assert.include(body, 'rmarkdown::render(input = \\"/erc/main.Rmd\\"');
-        done();
-      });
-    });
-
-    it('should have the manifest build log in the compendium metadata files', (done) => {
-      request(global.test_host + '/api/v1/compendium/' + compendium_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children
-          .find((element) => { return element.name === '.erc' })
-          .children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/compendium/' + compendium_id + '/data/.erc/generate_manifest.log');
-        done();
-      });
-    });
-
-    it('should complete build, execute, and cleanup', function (done) {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        assert.propertyVal(response.steps.image_build, 'status', 'success');
-        assert.propertyVal(response.steps.image_execute, 'status', 'success');
-        assert.propertyVal(response.steps.cleanup, 'status', 'success');
-        done();
-      });
-    });
-
-    it('should have the image build log in the compendium metadata files', (done) => {
-      request(global.test_host + '/api/v1/compendium/' + compendium_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children
-          .find((element) => { return element.name === '.erc' })
-          .children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/compendium/' + compendium_id + '/data/.erc/image_build.log');
-        done();
-      });
-    });
-
-  });
-
-  describe('Manifest generation for workspace minimal-script', () => {
-    let job_id = '';
-    let compendium_id = '';
-
-    before(function (done) {
-      this.timeout(90000);
-      createCompendiumPostRequest('./test/workspace/minimal-script', cookie_o2r, 'workspace', (req) => {
-        request(req, (err, res, body) => {
-          compendium_id = JSON.parse(body).id;
-          publishCandidate(compendium_id, cookie_o2r, () => {
-            startJob(compendium_id, id => {
-              job_id = id;
-              waitForJob(job_id, (finalStatus) => {
-                done();
-              });
-            });
-          });
-        });
-      });
-    });
-
-    it('should skip validation steps because it is a workspace', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        assert.propertyVal(response.steps.validate_bag, 'status', 'skipped');
-        assert.propertyVal(response.steps.validate_compendium, 'status', 'success');
-        done();
-      });
-    });
-
-    it('should complete step "generate_manifest"', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-        assert.propertyVal(response.steps.generate_manifest, 'status', 'success');
-        done();
-      });
-    });
-
-    it('show have the manifest file in the job files', (done) => {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/job/' + job_id + '/data/Dockerfile');
-        done();
-      });
-    });
-
-    it('should have the manifest file in the compendium files', (done) => {
-      request(global.test_host + '/api/v1/compendium/' + compendium_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        filePaths = response.files.children.map(elem => { return elem.path; });
-        assert.include(filePaths, '/api/v1/compendium/' + compendium_id + '/data/Dockerfile');
-        done();
-      });
-    });
-
-    it('should have the expected content in the manifest', function (done) {
-      request(global.test_host_transporter + '/api/v1/job/' + job_id + '/data/Dockerfile', (err, res, body) => {
-        assert.ifError(err);
-        assert.isNotObject(body, 'response is not JSON');
-        assert.notInclude(body, 'COPY', 'no COPY statement, because files are mounted');
-        assert.include(body, 'CMD ["R", "--vanilla", "-f", "main.R"]');
-        done();
-      });
-    });
-
-    it('should complete build, execute, and cleanup', function (done) {
-      request(global.test_host + '/api/v1/job/' + job_id, (err, res, body) => {
-        assert.ifError(err);
-        let response = JSON.parse(body);
-
-        assert.propertyVal(response.steps.image_build, 'status', 'success');
-        assert.propertyVal(response.steps.image_execute, 'status', 'success');
-        assert.propertyVal(response.steps.cleanup, 'status', 'success');
-        done();
-      });
-    });
-
-  });
-
   describe('EXECUTION step_image_prepare', () => {
     let job_id = '';
