+++ test/setup.js
@@ -25,6 +25,7 @@ const async = require('async');
 const tar = require('tar');
 const path = require('path');
 const fs = require('fs');
+const colors = require('colors');
 const debug = require('debug')('test:setup');
 
 // test parameters for local session authentication directly via fixed database entries
@@ -244,7 +245,8 @@ before(function (done) {
     pullBaseImageForManifestGeneration = (cb) => {
         docker.pull(config.containerit.baseImage, function (err, stream) {
             if (err) {
-                cb(err);
+                debug('ERROR pulling image, this may result in errors if image is not cached locally: %o'.yellow, err);
+                cb(null, err);
             } else {
                 function onFinished(err, output) {
                     if (err) cb(err);
@@ -296,7 +298,7 @@ before(function (done) {
     ],
         function (err, results) {
             if (err) {
-                debug('Error during test setup: %o', err);
+                debug('Error during test setup: %o'.red, err);
                 process.exit(1);
             } else {
                 debug('Test setup result: %o', results);
