+++ test/workspace/minimal-rmd-data/main.Rmd
@@ -2,7 +2,7 @@
 title: "Capacity of container ships in seaborne trade from 1980 to 2016 (in million dwt)*"
 author:
     - name: "Daniel Nüst"
-    - affiliation: o2r team
+      affiliation: o2r team
 date: "2017"
 output: html_document
 abstract: |
